refactor payment processing helpers
